+++ NEWS.md
@@ -3,6 +3,8 @@
 * Fixed bug in CSV logger callback that was saving the logs as a space delimited file (#52, @mattwarkentin).
 * Fixed bug in the length of the progress bar for the validation dataset (#52, @mattwarkentin).
 * `ctx$data` now refers to the current in use `data` instead of always refering to `ctx$train_data`. (#54)
+* Allow users to provide the minimum and maximum number of epochs when calling `fit.luz_module_generator()`. Removed `ctx$epochs` from context object and replaced it with `ctx$min_epochs` and `ctx$max_epochs` (#53, @mattwarkentin)
+* Early stopping will now only occur if the minimum number of training epochs has been met (#53, @mattwarkentin)
 
 # luz 0.1.0